Renovated editors: refactor directory structure
