+++ js/renovation/component_wrapper/editors/__tests__/check_box.test.tsx
@@ -1,6 +1,6 @@
 import CheckBox from '../check_box';
 
-import { addAttributes, getAriaName } from '../utils/utils';
+import { addAttributes, getAriaName } from '../../utils/utils';
 import Mock = jest.Mock;
 
 jest.mock('../utils/utils');
