+++ js/renovation/component_wrapper/editors/__tests__/editor.test.tsx
