+++ js/renovation/component_wrapper/editors/check_box.ts
@@ -1,8 +1,8 @@
 // eslint-disable-next-line import/named
-import { dxElementWrapper } from '../../core/renderer';
-import Editor from './common/editor';
-import { Option } from './common/types';
-import { addAttributes, getAriaName } from './utils/utils';
+import { dxElementWrapper } from '../../../core/renderer';
+import Editor from './editor';
+import { Option } from '../common/types';
+import { addAttributes, getAriaName } from '../utils/utils';
 
 export default class CheckBox extends Editor {
   // eslint-disable-next-line class-methods-use-this
