+++ js/renovation/ui/drop_down_editors/__tests__/date_box.test.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import LegacyDateBox from '../../../ui/date_box';
+import LegacyDateBox from '../../../../ui/date_box';
 import { viewFunction as DateBoxView, DateBoxProps, DateBox } from '../date_box';
-import { DomComponentWrapper } from '../common/dom_component_wrapper';
+import { DomComponentWrapper } from '../../common/dom_component_wrapper';
 
 jest.mock('../../../ui/date_box', () => jest.fn());
