+++ js/renovation/ui/editors/number_box.tsx
@@ -2,10 +2,10 @@ import {
   Component, ComponentBindings, JSXComponent, OneWay, Event, TwoWay, React,
 } from '@devextreme-generator/declarations';
 /* eslint-disable-next-line import/named */
-import LegacyNumberBox from '../../ui/number_box';
-import { DomComponentWrapper } from './common/dom_component_wrapper';
-import { EventCallback } from './common/event_callback.d';
-import { BaseWidgetProps } from './common/base_props';
+import LegacyNumberBox from '../../../ui/number_box';
+import { DomComponentWrapper } from '../common/dom_component_wrapper';
+import { EventCallback } from '../common/event_callback';
+import { BaseWidgetProps } from '../common/base_props';
 
 export const viewFunction = ({
   props,
