+++ js/renovation/ui/pager/page_size/small.tsx
@@ -2,7 +2,7 @@ import {
   ComponentBindings, JSXComponent, OneWay, InternalState, Effect, Component, Ref, RefObject,
 } from '@devextreme-generator/declarations';
 
-import { SelectBox } from '../../select_box';
+import { SelectBox } from '../../drop_down_editors/select_box';
 import { calculateValuesFittedWidth } from '../utils/calculate_values_fitted_width';
 import { FullPageSize } from '../common/types.d';
 import { getElementMinWidth } from '../utils/get_element_width';
