+++ js/renovation/ui/scheduler/appointment_tooltip/appointment_list.tsx
@@ -8,7 +8,7 @@ import {
   JSXTemplate,
 } from '@devextreme-generator/declarations';
 import noop from '../../../utils/noop';
-import { List } from '../../list';
+import { List } from '../../editors/list';
 import { TooltipItemLayout } from './item_layout';
 import {
   GetTextAndFormatDateFn,
